Rename singletone directory and update imports
